Plot A Tocky Network

Plot A Tocky Network

Usage

PlotTockyNetwork(
  x,
  reduction = "CCA",
  select_variable = FALSE,
  edge_scale = NULL,
  mds = FALSE,
  reflect = NULL,
  log2fc = FALSE,
  p_adjust = NULL
)

Arguments

x

A TockyPrepData after running the function TockyNetwork.

reduction

Choose whether to use PCA or CCA as a reduction method.

select_variable

Whether to choose a variable (marker expression). The default is FALSE and produces Tocky Angle and Intensity.

edge_scale

A scale factor for edge thickness.

mds

If Multidimensional Scaling is used for constructing network graph.

reflect

If Multidimensional Scaling is used, graph can be refelected by either 'horizontal' or 'vertical'.

log2fc

Logical. If TRUE, the colours of nodes are determined by log2 fold change between the two groups.

p_adjust

If p_adjust is "none", p-value adjustment is not used. Note that p-values from two-group comparisons are stored within the TockyPrepData object and have been calculated by 'PlotClusterPercentage'.

Value

Network plot is produced, showing clusters as vertices and distance between clusters as edges.

Examples

if (FALSE) { # \dontrun{
PlotTockyNetwork(x)
} # }