Perform Principal Component Analysis for TockyPrepData.

Perform Principal Component Analysis for TockyPrepData.

Usage

TockyPCA(
  x,
  variables = NULL,
  marker_neg_gate = TRUE,
  cleaning = FALSE,
  select = TRUE,
  Timer = FALSE
)

Arguments

x

A TockyPrepData after running the function Tocky.

variables

Variables (markers) for PCA analysis. If NULL, variables are to be chosed interactively.

marker_neg_gate

Whether autofluorescence values are all considered to be zero. This approach is recommended and requires DefineNegative.

cleaning

Whether data cleaning is performed to remove cells with a zero value for a marker. This is not recommended unless you have a reason that you cannot use DefineNegative to collapse negative data.

select

Whether to interactively select markers to be processed. If FALSE, all the log-transformed markers apart from Timer fluorescence (stored in the TockyPrepData) will be used.

Timer

Whether Timer data (normalised or Angle/Intensity) are used.

Examples

if (FALSE) { # \dontrun{
x <- TockyPCA(x, variables = NULL, cleaning = FALSE, marker_neg_gate = TRUE)
} # }